requirements.txt
absl-py==0.12.0
astunparse==1.6.3
attrs==21.2.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Cython==0.29.24
dill==0.3.4
distlib==0.3.4
dm-tree==0.1.6
filelock==3.4.2
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
flatbuffers==2.0
fonttools==4.28.3
future==0.18.2
gast==0.4.0
google-api-core==2.2.2
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpcio==1.42.0
gunicorn==20.1.0
h5py==3.6.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
lvis==0.5.3
makefile==1.1.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.0
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
Pillow==8.4.0
platformdirs==2.4.1
portalocker==2.3.2
promise==2.3
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
pywin32==302
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
sacrebleu==2.0.0
scikit-learn==1.0.1
scipy==1.7.3
sentencepiece==0.1.96
seqeval==1.2.2
setuptools-scm==6.3.2
six==1.16.0
SQLAlchemy==1.4.31
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-addons==0.15.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.22.0
tensorflow-metadata==1.5.0
tensorflow-model-optimization==0.7.0
tensorflow-text==2.7.3
termcolor==1.1.0
text-unidecode==1.3
tf-slim==1.1.0
threadpoolctl==3.0.0
tomli==1.2.2
tqdm==4.62.3
typeguard==2.13.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.13.1
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
